+++ src/main/java/janggi/domain/movement/RuleOfCannon.java
@@ -2,7 +2,7 @@ package janggi.domain.movement;
 
 import static janggi.domain.Position.MAXIMUM_ROW;
 
-import janggi.domain.PieceType;
+import janggi.domain.piece.PieceType;
 import janggi.domain.Position;
 import janggi.domain.board.BoardMediator;
 import janggi.domain.piece.Piece;
