+++ src/main/java/janggi/domain/piece/Chariot.java
@@ -3,7 +3,6 @@ package janggi.domain.piece;
 import static janggi.domain.Position.MAXIMUM_COLUMN;
 import static janggi.domain.Position.MAXIMUM_ROW;
 
-import janggi.domain.PieceType;
 import janggi.domain.Position;
 import janggi.domain.board.BoardMediator;
 import janggi.domain.movement.Direction;
