+++ src/main/java/janggi/domain/piece/PieceType.java
@@ -1,13 +1,5 @@
-package janggi.domain;
+package janggi.domain.piece;
 
-import janggi.domain.piece.Cannon;
-import janggi.domain.piece.Chariot;
-import janggi.domain.piece.Elephant;
-import janggi.domain.piece.General;
-import janggi.domain.piece.Guard;
-import janggi.domain.piece.Horse;
-import janggi.domain.piece.Piece;
-import janggi.domain.piece.Soldier;
 import janggi.domain.team.TeamType;
 import java.util.function.Function;
