+++ src/test/java/janggi/domain/board/BoardGeneratorTest.java
@@ -1,9 +1,8 @@
-package janggi.domain;
+package janggi.domain.board;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import janggi.domain.board.Board;
-import janggi.domain.board.BoardGenerator;
+import janggi.domain.setup.InnerElephantSetupPolicy;
 import janggi.domain.team.BlueTeam;
 import janggi.domain.team.RedTeam;
 import org.assertj.core.api.InstanceOfAssertFactories;
