+++ src/test/java/janggi/domain/movement/DirectionTest.java
@@ -1,9 +1,8 @@
-package janggi.domain;
+package janggi.domain.movement;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.junit.jupiter.api.Assertions.*;
 
-import janggi.domain.movement.Direction;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
