+++ src/test/java/janggi/domain/piece/GuardTest.java
@@ -1,13 +1,11 @@
-package janggi.domain;
+package janggi.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import janggi.domain.Position;
 import janggi.domain.board.Board;
 import janggi.domain.board.BoardMediator;
 import janggi.domain.board.BoardMediatorImpl;
-import janggi.domain.piece.Horse;
-import janggi.domain.piece.Piece;
-import janggi.domain.piece.Soldier;
 import janggi.domain.team.TeamType;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -17,13 +15,13 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-public class HorseTest {
+public class GuardTest {
 
     @Nested
     @DisplayName("이동 가능한 위치 계산 테스트")
     class CalculateMovablePositions {
 
-        static Piece horse;
+        static Piece guard;
         static Piece enemy1;
         static Piece enemy2;
         static Piece enemy3;
@@ -36,7 +34,7 @@ public class HorseTest {
 
         @BeforeEach
         void setUp() {
-            horse = new Horse(TeamType.RED);
+            guard = new Guard(TeamType.RED);
             enemy1 = new Soldier(TeamType.BLUE);
             enemy3 = new Soldier(TeamType.BLUE);
             enemy4 = new Soldier(TeamType.BLUE);
@@ -49,53 +47,59 @@ public class HorseTest {
         }
 
         @Test
-        @DisplayName("마는 기물을 뛰어넘을 수 없다.")
+        @DisplayName("사는 기물을 뛰어넘을 수 없다.")
         void test1() {
-            positionPieceMap.put(Position.valueOf(6, 4), horse);
-            positionPieceMap.put(Position.valueOf(6, 3), ally1);
-            positionPieceMap.put(Position.valueOf(5, 4), enemy1);
-            positionPieceMap.put(Position.valueOf(7, 4), ally2);
-            positionPieceMap.put(Position.valueOf(6, 5), ally3);
-            List<Position> expected = List.of();
+            positionPieceMap.put(Position.valueOf(6, 4), guard);
+            positionPieceMap.put(Position.valueOf(5, 3), ally1);
+            positionPieceMap.put(Position.valueOf(5, 4), ally2);
+            positionPieceMap.put(Position.valueOf(5, 5), ally3);
+            positionPieceMap.put(Position.valueOf(6, 3), ally4);
+            positionPieceMap.put(Position.valueOf(6, 5), enemy1);
+            positionPieceMap.put(Position.valueOf(7, 3), enemy2);
+            positionPieceMap.put(Position.valueOf(7, 4), enemy3);
+            positionPieceMap.put(Position.valueOf(7, 5), enemy4);
+
+            List<Position> expected = List.of(Position.valueOf(6, 5), Position.valueOf(7, 3),
+                Position.valueOf(7, 4), Position.valueOf(7, 5));
 
             Board board = new Board(positionPieceMap);
             BoardMediator boardMediator = new BoardMediatorImpl(board);
-            List<Position> actual = horse.calculateMovablePositions(Position.valueOf(6, 4),
+            List<Position> actual = guard.calculateMovablePositions(Position.valueOf(6, 4),
                 boardMediator);
 
             assertThat(actual).hasSameElementsAs(expected);
         }
 
         @Test
-        @DisplayName("마는 직선 한 칸, 대각선 한 칸을 가서 기물을 잡을 수 있다.")
+        @DisplayName("사는 모든 방향 중 한 칸을 가서 기물을 잡을 수 있다.")
         void test2() {
-            positionPieceMap.put(Position.valueOf(6, 4), horse);
-            positionPieceMap.put(Position.valueOf(5, 6), enemy1);
-            positionPieceMap.put(Position.valueOf(7, 2), enemy2);
-            List<Position> expected = List.of(Position.valueOf(4, 3), Position.valueOf(4, 5),
-                Position.valueOf(5, 6), Position.valueOf(7, 6),
-                Position.valueOf(8, 3), Position.valueOf(8, 5), Position.valueOf(7, 2),
-                Position.valueOf(5, 2));
+            positionPieceMap.put(Position.valueOf(6, 4), guard);
+            positionPieceMap.put(Position.valueOf(6, 5), enemy1);
+            positionPieceMap.put(Position.valueOf(7, 3), enemy2);
+            List<Position> expected = List.of(Position.valueOf(5, 3), Position.valueOf(5, 4),
+                Position.valueOf(5, 5), Position.valueOf(6, 3), Position.valueOf(6, 5), Position.valueOf(7, 3),
+                Position.valueOf(7, 4), Position.valueOf(7, 5));
 
             Board board = new Board(positionPieceMap);
             BoardMediator boardMediator = new BoardMediatorImpl(board);
-            List<Position> actual = horse.calculateMovablePositions(Position.valueOf(6, 4),
+            List<Position> actual = guard.calculateMovablePositions(Position.valueOf(6, 4),
                 boardMediator);
 
             assertThat(actual).hasSameElementsAs(expected);
         }
 
         @Test
-        @DisplayName("마는 장기판 밖으로 이동할 수 없다.")
+        @DisplayName("사는 장기판 밖으로 이동할 수 없다.")
         void test3() {
-            positionPieceMap.put(Position.valueOf(1, 1), horse);
+            positionPieceMap.put(Position.valueOf(1, 1), guard);
             positionPieceMap.put(Position.valueOf(1, 2), ally1);
-            positionPieceMap.put(Position.valueOf(2, 1), ally3);
+            positionPieceMap.put(Position.valueOf(2, 1), ally2);
+            positionPieceMap.put(Position.valueOf(2, 2), ally3);
             List<Position> expected = List.of();
 
             Board board = new Board(positionPieceMap);
             BoardMediator boardMediator = new BoardMediatorImpl(board);
-            List<Position> actual = horse.calculateMovablePositions(Position.valueOf(1, 1),
+            List<Position> actual = guard.calculateMovablePositions(Position.valueOf(1, 1),
                 boardMediator);
 
             assertThat(actual).hasSameElementsAs(expected);
