+++ src/test/java/janggi/domain/piece/RedSoldierTest.java
@@ -1,12 +1,11 @@
-package janggi.domain;
+package janggi.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import janggi.domain.Position;
 import janggi.domain.board.Board;
 import janggi.domain.board.BoardMediator;
 import janggi.domain.board.BoardMediatorImpl;
-import janggi.domain.piece.Piece;
-import janggi.domain.piece.Soldier;
 import janggi.domain.team.TeamType;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -16,13 +15,13 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-public class BlueSoldierTest {
+public class RedSoldierTest {
 
     @Nested
-    @DisplayName("Blue Soldier 이동 가능한 위치 계산 테스트")
+    @DisplayName("Red Soldier 이동 가능한 위치 계산 테스트")
     class CalculateMovablePositions {
 
-        static Piece blueSoldier;
+        static Piece redSoldier;
         static Piece enemy1;
         static Piece enemy2;
         static Piece ally1;
@@ -31,23 +30,23 @@ public class BlueSoldierTest {
 
         @BeforeEach
         void setUp() {
-            blueSoldier = new Soldier(TeamType.BLUE);
-            enemy1 = new Soldier(TeamType.RED);
-            enemy2 = new Soldier(TeamType.RED);
-            ally1 = new Soldier(TeamType.BLUE);
-            ally2 = new Soldier(TeamType.BLUE);
+            redSoldier = new Soldier(TeamType.RED);
+            enemy1 = new Soldier(TeamType.BLUE);
+            enemy2 = new Soldier(TeamType.BLUE);
+            ally1 = new Soldier(TeamType.RED);
+            ally2 = new Soldier(TeamType.RED);
             positionPieceMap = new LinkedHashMap<>();
         }
 
         @Test
-        @DisplayName("청졸은 앞, 좌, 우로 이동할 수 있고 적이 있으면 잡을 수 있다.")
+        @DisplayName("홍졸은 앞, 좌, 우로 이동할 수 있고 적이 있으면 잡을 수 있다.")
         void test1() {
-            positionPieceMap.put(Position.valueOf(6, 4), blueSoldier);
-            positionPieceMap.put(Position.valueOf(5, 4), enemy1);
+            positionPieceMap.put(Position.valueOf(6, 4), redSoldier);
+            positionPieceMap.put(Position.valueOf(7, 4), enemy1);
             positionPieceMap.put(Position.valueOf(6, 3), enemy2);
 
             List<Position> expected = List.of(
-                    Position.valueOf(5, 4),
+                    Position.valueOf(7, 4),
                     Position.valueOf(6, 3),
                     Position.valueOf(6, 5)
             );
@@ -55,7 +54,7 @@ public class BlueSoldierTest {
             Board board = new Board(positionPieceMap);
             BoardMediator boardMediator = new BoardMediatorImpl(board);
 
-            List<Position> actual = blueSoldier.calculateMovablePositions(
+            List<Position> actual = redSoldier.calculateMovablePositions(
                     Position.valueOf(6, 4), boardMediator
             );
 
@@ -63,10 +62,10 @@ public class BlueSoldierTest {
         }
 
         @Test
-        @DisplayName("청졸은 아군이 있는 위치로 이동할 수 없다.")
+        @DisplayName("홍졸은 아군이 있는 위치로 이동할 수 없다.")
         void test2() {
-            positionPieceMap.put(Position.valueOf(6, 4), blueSoldier);
-            positionPieceMap.put(Position.valueOf(5, 4), ally1);
+            positionPieceMap.put(Position.valueOf(6, 4), redSoldier);
+            positionPieceMap.put(Position.valueOf(7, 4), ally1);
             positionPieceMap.put(Position.valueOf(6, 5), ally2);
 
             List<Position> expected = List.of(
@@ -76,7 +75,7 @@ public class BlueSoldierTest {
             Board board = new Board(positionPieceMap);
             BoardMediator boardMediator = new BoardMediatorImpl(board);
 
-            List<Position> actual = blueSoldier.calculateMovablePositions(
+            List<Position> actual = redSoldier.calculateMovablePositions(
                     Position.valueOf(6, 4), boardMediator
             );
 
@@ -84,18 +83,18 @@ public class BlueSoldierTest {
         }
 
         @Test
-        @DisplayName("청졸은 장기판 밖으로 이동할 수 없다.")
+        @DisplayName("홍졸은 장기판 밖으로 이동할 수 없다.")
         void test3() {
-            positionPieceMap.put(Position.valueOf(1, 1), blueSoldier);
-            positionPieceMap.put(Position.valueOf(1, 2), ally1);
+            positionPieceMap.put(Position.valueOf(10, 1), redSoldier);
+            positionPieceMap.put(Position.valueOf(10, 2), ally1);
 
             List<Position> expected = List.of();
 
             Board board = new Board(positionPieceMap);
             BoardMediator boardMediator = new BoardMediatorImpl(board);
 
-            List<Position> actual = blueSoldier.calculateMovablePositions(
-                    Position.valueOf(1, 1), boardMediator
+            List<Position> actual = redSoldier.calculateMovablePositions(
+                    Position.valueOf(10, 1), boardMediator
             );
 
             assertThat(actual).hasSameElementsAs(expected);
